+++ Specs/DataSources/BillboardVisualizerSpec.js
@@ -12,6 +12,7 @@ defineSuite([
         'DataSources/BillboardGraphics',
         'DataSources/BoundingSphereState',
         'DataSources/ConstantProperty',
+        'DataSources/EntityCluster',
         'DataSources/EntityCollection',
         'Scene/BillboardCollection',
         'Scene/HeightReference',
@@ -33,6 +34,7 @@ defineSuite([
         BillboardGraphics,
         BoundingSphereState,
         ConstantProperty,
+        EntityCluster,
         EntityCollection,
         BillboardCollection,
         HeightReference,
@@ -44,6 +46,7 @@ defineSuite([
     'use strict';
 
     var scene;
+    var entityCluster;
     var visualizer;
 
     beforeAll(function() {
@@ -55,11 +58,17 @@ defineSuite([
         scene.destroyForSpecs();
     });
 
-    afterEach(function() {
-        visualizer = visualizer && visualizer.destroy();
+    beforeEach(function() {
+        entityCluster = new EntityCluster();
+        entityCluster._initialize(scene);
     });
 
-    it('constructor throws if no scene is passed.', function() {
+    afterEach(function() {
+        visualizer = visualizer && visualizer.destroy();
+        entityCluster.destroy();
+    });
+
+    it('constructor throws if no entityCluster is passed.', function() {
         expect(function() {
             return new BillboardVisualizer();
         }).toThrowDeveloperError();
@@ -67,7 +76,7 @@ defineSuite([
 
     it('update throws if no time specified.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
         expect(function() {
             visualizer.update();
         }).toThrowDeveloperError();
@@ -75,7 +84,7 @@ defineSuite([
 
     it('isDestroy returns false until destroyed.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
         expect(visualizer.isDestroyed()).toEqual(false);
         visualizer.destroy();
         expect(visualizer.isDestroyed()).toEqual(true);
@@ -84,25 +93,25 @@ defineSuite([
 
     it('removes the listener from the entity collection when destroyed', function() {
         var entityCollection = new EntityCollection();
-        var visualizer = new BillboardVisualizer(scene, entityCollection);
+        var visualizer = new BillboardVisualizer(entityCluster, entityCollection);
         expect(entityCollection.collectionChanged.numberOfListeners).toEqual(1);
-        visualizer = visualizer.destroy();
+        visualizer.destroy();
         expect(entityCollection.collectionChanged.numberOfListeners).toEqual(0);
     });
 
     it('object with no billboard does not create a billboard.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         testObject.position = new ConstantProperty(new Cartesian3(1234, 5678, 9101112));
         visualizer.update(JulianDate.now());
-        expect(scene.primitives.length).toEqual(0);
+        expect(entityCluster._billboardCollection).not.toBeDefined();
     });
 
     it('object with no position does not create a billboard.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         var billboard = testObject.billboard = new BillboardGraphics();
@@ -110,12 +119,12 @@ defineSuite([
         billboard.image = new ConstantProperty('Data/Images/Blue.png');
 
         visualizer.update(JulianDate.now());
-        expect(scene.primitives.length).toEqual(0);
+        expect(entityCluster._billboardCollection).not.toBeDefined();
     });
 
     it('object with no image does not create a billboard.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         testObject.position = new ConstantProperty(new Cartesian3(1234, 5678, 9101112));
@@ -123,12 +132,12 @@ defineSuite([
         billboard.show = new ConstantProperty(true);
 
         visualizer.update(JulianDate.now());
-        expect(scene.primitives.length).toEqual(0);
+        expect(entityCluster._billboardCollection).not.toBeDefined();
     });
 
     it('A BillboardGraphics causes a Billboard to be created and updated.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
 
@@ -159,7 +168,7 @@ defineSuite([
 
         visualizer.update(time);
 
-        var billboardCollection = scene.primitives.get(0);
+        var billboardCollection = entityCluster._billboardCollection;
         expect(billboardCollection.length).toEqual(1);
 
         bb = billboardCollection.get(0);
@@ -198,7 +207,7 @@ defineSuite([
     it('Reuses primitives when hiding one and showing another', function() {
         var time = JulianDate.now();
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         testObject.position = new ConstantProperty(new Cartesian3(1234, 5678, 9101112));
@@ -208,7 +217,7 @@ defineSuite([
 
         visualizer.update(time);
 
-        var billboardCollection = scene.primitives.get(0);
+        var billboardCollection = entityCluster._billboardCollection;
         expect(billboardCollection.length).toEqual(1);
 
         testObject.billboard.show = new ConstantProperty(false);
@@ -229,7 +238,7 @@ defineSuite([
 
     it('clear hides billboards.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
 
@@ -241,7 +250,7 @@ defineSuite([
         billboard.image = new ConstantProperty('Data/Images/Blue.png');
         visualizer.update(time);
 
-        var billboardCollection = scene.primitives.get(0);
+        var billboardCollection = entityCluster._billboardCollection;
         expect(billboardCollection.length).toEqual(1);
         var bb = billboardCollection.get(0);
 
@@ -259,7 +268,7 @@ defineSuite([
 
     it('Visualizer sets entity property.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         var time = JulianDate.now();
@@ -270,7 +279,7 @@ defineSuite([
         billboard.image = new ConstantProperty('Data/Images/Blue.png');
         visualizer.update(time);
 
-        var billboardCollection = scene.primitives.get(0);
+        var billboardCollection = entityCluster._billboardCollection;
         expect(billboardCollection.length).toEqual(1);
         var bb = billboardCollection.get(0);
         expect(bb.id).toEqual(testObject);
@@ -278,7 +287,7 @@ defineSuite([
 
     it('Computes bounding sphere.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         var time = JulianDate.now();
@@ -300,7 +309,7 @@ defineSuite([
     it('Fails bounding sphere for entity without billboard.', function() {
         var entityCollection = new EntityCollection();
         var testObject = entityCollection.getOrCreateEntity('test');
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
         visualizer.update(JulianDate.now());
         var result = new BoundingSphere();
         var state = visualizer.getBoundingSphere(testObject, result);
@@ -309,7 +318,7 @@ defineSuite([
 
     it('Compute bounding sphere throws without entity.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
         var result = new BoundingSphere();
         expect(function() {
             visualizer.getBoundingSphere(undefined, result);
@@ -319,7 +328,7 @@ defineSuite([
     it('Compute bounding sphere throws without result.', function() {
         var entityCollection = new EntityCollection();
         var testObject = entityCollection.getOrCreateEntity('test');
-        visualizer = new BillboardVisualizer(scene, entityCollection);
+        visualizer = new BillboardVisualizer(entityCluster, entityCollection);
         expect(function() {
             visualizer.getBoundingSphere(testObject, undefined);
         }).toThrowDeveloperError();