+++ Specs/DataSources/PointVisualizerSpec.js
@@ -12,6 +12,7 @@ defineSuite([
         'Core/NearFarScalar',
         'DataSources/BoundingSphereState',
         'DataSources/ConstantProperty',
+        'DataSources/EntityCluster',
         'DataSources/EntityCollection',
         'DataSources/PointGraphics',
         'Scene/BillboardCollection',
@@ -31,6 +32,7 @@ defineSuite([
         NearFarScalar,
         BoundingSphereState,
         ConstantProperty,
+        EntityCluster,
         EntityCollection,
         PointGraphics,
         BillboardCollection,
@@ -40,6 +42,7 @@ defineSuite([
     'use strict';
 
     var scene;
+    var entityCluster;
     var visualizer;
 
     beforeAll(function() {
@@ -73,8 +76,14 @@ defineSuite([
         scene.destroyForSpecs();
     });
 
+    beforeEach(function() {
+        entityCluster = new EntityCluster();
+        entityCluster._initialize(scene);
+    });
+
     afterEach(function() {
         visualizer = visualizer && visualizer.destroy();
+        entityCluster.destroy();
     });
 
     it('constructor throws if no scene is passed.', function() {
@@ -86,13 +95,13 @@ defineSuite([
 
     it('constructor throws if no entityCollection is passed.', function() {
         expect(function() {
-            return new PointVisualizer(scene, undefined);
+            return new PointVisualizer(entityCluster, undefined);
         }).toThrowDeveloperError();
     });
 
     it('update throws if no time specified.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
         expect(function() {
             visualizer.update();
         }).toThrowDeveloperError();
@@ -100,7 +109,7 @@ defineSuite([
 
     it('isDestroy returns false until destroyed.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
         expect(visualizer.isDestroyed()).toEqual(false);
         visualizer.destroy();
         expect(visualizer.isDestroyed()).toEqual(true);
@@ -109,7 +118,7 @@ defineSuite([
 
     it('removes the listener from the entity collection when destroyed', function() {
         var entityCollection = new EntityCollection();
-        var visualizer = new PointVisualizer(scene, entityCollection);
+        var visualizer = new PointVisualizer(entityCluster, entityCollection);
         expect(entityCollection.collectionChanged.numberOfListeners).toEqual(1);
         visualizer = visualizer.destroy();
         expect(entityCollection.collectionChanged.numberOfListeners).toEqual(0);
@@ -117,7 +126,7 @@ defineSuite([
 
     it('object with no point does not create a pointPrimitive.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         testObject.position = new ConstantProperty(new Cartesian3(1234, 5678, 9101112));
@@ -127,7 +136,7 @@ defineSuite([
 
     it('object with no position does not create a pointPrimitive.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         var point = testObject.point = new PointGraphics();
@@ -141,7 +150,7 @@ defineSuite([
         var time = JulianDate.now();
 
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var entity = entityCollection.add({
             position : new Cartesian3(1234, 5678, 9101112),
@@ -159,7 +168,7 @@ defineSuite([
 
         visualizer.update(time);
 
-        var pointPrimitiveCollection = scene.primitives.get(0);
+        var pointPrimitiveCollection = entityCluster._pointCollection;
         expect(pointPrimitiveCollection instanceof PointPrimitiveCollection).toBe(true);
         expect(pointPrimitiveCollection.length).toEqual(1);
         var pointPrimitive = pointPrimitiveCollection.get(0);
@@ -198,7 +207,7 @@ defineSuite([
         var time = JulianDate.now();
 
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var entity = entityCollection.add({
             position : new Cartesian3(1234, 5678, 9101112),
@@ -217,7 +226,7 @@ defineSuite([
 
         visualizer.update(time);
 
-        var billboardCollection = scene.primitives.get(0);
+        var billboardCollection = entityCluster._billboardCollection;
         expect(billboardCollection instanceof BillboardCollection).toBe(true);
         expect(billboardCollection.length).toEqual(1);
         var billboard = billboardCollection.get(0);
@@ -255,7 +264,7 @@ defineSuite([
     it('Reuses primitives when hiding one and showing another', function() {
         var time = JulianDate.now();
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         testObject.position = new ConstantProperty(new Cartesian3(1234, 5678, 9101112));
@@ -264,7 +273,7 @@ defineSuite([
 
         visualizer.update(time);
 
-        var pointPrimitiveCollection = scene.primitives.get(0);
+        var pointPrimitiveCollection = entityCluster._pointCollection;
         expect(pointPrimitiveCollection.length).toEqual(1);
 
         testObject.point.show = new ConstantProperty(false);
@@ -284,7 +293,7 @@ defineSuite([
 
     it('clear hides pointPrimitives.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
         var testObject = entityCollection.getOrCreateEntity('test');
         var time = JulianDate.now();
 
@@ -293,7 +302,7 @@ defineSuite([
         point.show = new ConstantProperty(true);
         visualizer.update(time);
 
-        var pointPrimitiveCollection = scene.primitives.get(0);
+        var pointPrimitiveCollection = entityCluster._pointCollection;
         expect(pointPrimitiveCollection.length).toEqual(1);
         var bb = pointPrimitiveCollection.get(0);
 
@@ -307,7 +316,7 @@ defineSuite([
 
     it('Visualizer sets entity property.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         var time = JulianDate.now();
@@ -318,7 +327,7 @@ defineSuite([
 
         visualizer.update(time);
 
-        var pointPrimitiveCollection = scene.primitives.get(0);
+        var pointPrimitiveCollection = entityCluster._pointCollection;
         expect(pointPrimitiveCollection.length).toEqual(1);
         var bb = pointPrimitiveCollection.get(0);
         expect(bb.id).toEqual(testObject);
@@ -326,7 +335,7 @@ defineSuite([
 
     it('Computes bounding sphere.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
 
         var testObject = entityCollection.getOrCreateEntity('test');
         var time = JulianDate.now();
@@ -348,7 +357,7 @@ defineSuite([
     it('Fails bounding sphere for entity without pointPrimitive.', function() {
         var entityCollection = new EntityCollection();
         var testObject = entityCollection.getOrCreateEntity('test');
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
         visualizer.update(JulianDate.now());
         var result = new BoundingSphere();
         var state = visualizer.getBoundingSphere(testObject, result);
@@ -357,7 +366,7 @@ defineSuite([
 
     it('Compute bounding sphere throws without entity.', function() {
         var entityCollection = new EntityCollection();
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
         var result = new BoundingSphere();
         expect(function() {
             visualizer.getBoundingSphere(undefined, result);
@@ -367,7 +376,7 @@ defineSuite([
     it('Compute bounding sphere throws without result.', function() {
         var entityCollection = new EntityCollection();
         var testObject = entityCollection.getOrCreateEntity('test');
-        visualizer = new PointVisualizer(scene, entityCollection);
+        visualizer = new PointVisualizer(entityCluster, entityCollection);
         expect(function() {
             visualizer.getBoundingSphere(testObject, undefined);
         }).toThrowDeveloperError();