+++ Specs/Scene/BillboardCollectionSpec.js
@@ -17,6 +17,7 @@ defineSuite([
         'Core/NearFarScalar',
         'Core/Rectangle',
         'Renderer/ContextLimits',
+        'Scene/Billboard',
         'Scene/HeightReference',
         'Scene/HorizontalOrigin',
         'Scene/OrthographicFrustum',
@@ -44,6 +45,7 @@ defineSuite([
         NearFarScalar,
         Rectangle,
         ContextLimits,
+        Billboard,
         HeightReference,
         HorizontalOrigin,
         OrthographicFrustum,
@@ -601,7 +603,7 @@ defineSuite([
         });
         billboards.add({
             position : new Cartesian3(-1.0, 0.0, 0.0),
-            image : blueImage
+            image : largeBlueImage
         });
 
         expect(scene.renderForSpecs()).toEqual([0, 255, 0, 255]);
@@ -630,7 +632,7 @@ defineSuite([
 
         billboards.add({
             position : new Cartesian3(1.0, 0.0, 0.0), // Closer to camera
-            image : blueImage
+            image : largeBlueImage
         });
 
         expect(scene.renderForSpecs()).toEqual([0, 0, 255, 255]);
@@ -643,7 +645,7 @@ defineSuite([
         });
         var blueBillboard = billboards.add({
             position : new Cartesian3(1.0, 0.0, 0.0), // Closer to camera
-            image : blueImage
+            image : largeBlueImage
         });
 
         expect(scene.renderForSpecs()).toEqual([0, 0, 255, 255]);
@@ -675,7 +677,7 @@ defineSuite([
         billboards.removeAll();
         billboards.add({
             position : Cartesian3.ZERO,
-            image : blueImage
+            image : largeBlueImage
         });
 
         expect(scene.renderForSpecs()).toEqual([0, 0, 255, 255]);
@@ -702,7 +704,7 @@ defineSuite([
         var blueBillboard = billboards.add({
             show : false,
             position : Cartesian3.ZERO,
-            image : blueImage
+            image : largeBlueImage
         });
 
         expect(scene.renderForSpecs()).toEqual([0, 255, 0, 255]);
@@ -751,7 +753,7 @@ defineSuite([
 
         expect(scene.renderForSpecs()).toEqual([0, 255, 0, 255]);
 
-        b.image = blueImage;
+        b.image = largeBlueImage;
         expect(scene.renderForSpecs()).toEqual([0, 0, 255, 255]);
     });
 
@@ -763,7 +765,7 @@ defineSuite([
 
         expect(scene.renderForSpecs()).toEqual([0, 255, 0, 255]);
 
-        b.setImage(blueImage.src, blueImage);
+        b.setImage(largeBlueImage.src, largeBlueImage);
         expect(scene.renderForSpecs()).toEqual([0, 0, 255, 255]);
     });
 
@@ -862,7 +864,7 @@ defineSuite([
 
         var b2 = billboards.add({
             position : new Cartesian3(1.0, 0.0, 0.0), // Closer to camera
-            image : blueImage
+            image : largeBlueImage
         });
         expect(scene.renderForSpecs()).toEqual([0, 0, 255, 255]);
 
@@ -1094,6 +1096,114 @@ defineSuite([
         expect(function() {
             b.computeScreenSpacePosition();
         }).toThrowDeveloperError();
+    });
+
+    it('computes screen space bounding box', function() {
+        var width = 10;
+        var height = 15;
+        var scale = 1.5;
+
+        var b = billboards.add({
+            width : width,
+            height : height,
+            scale : scale
+        });
+
+        var halfWidth = width * scale * 0.5;
+        var halfHeight = height * scale * 0.5;
+        width = width * scale;
+        height = height * scale;
+
+        var bbox = Billboard.getScreenSpaceBoundingBox(b, Cartesian2.ZERO);
+        expect(bbox.x).toEqual(-halfWidth);
+        expect(bbox.y).toEqual(-halfHeight);
+        expect(bbox.width).toEqual(width);
+        expect(bbox.height).toEqual(height);
+    });
+
+    it('computes screen space bounding box with result', function() {
+        var width = 10;
+        var height = 15;
+        var scale = 1.5;
+
+        var b = billboards.add({
+            width : width,
+            height : height,
+            scale : scale
+        });
+
+        var halfWidth = width * scale * 0.5;
+        var halfHeight = height * scale * 0.5;
+        width = width * scale;
+        height = height * scale;
+
+        var result = new BoundingRectangle();
+        var bbox = Billboard.getScreenSpaceBoundingBox(b, Cartesian2.ZERO, result);
+        expect(bbox.x).toEqual(-halfWidth);
+        expect(bbox.y).toEqual(-halfHeight);
+        expect(bbox.width).toEqual(width);
+        expect(bbox.height).toEqual(height);
+        expect(bbox).toBe(result);
+    });
+
+    it('computes screen space bounding box with vertical origin', function() {
+        var width = 10;
+        var height = 15;
+        var scale = 1.5;
+
+        var b = billboards.add({
+            width : width,
+            height : height,
+            scale : scale,
+            verticalOrigin : VerticalOrigin.BOTTOM
+        });
+
+        var halfWidth = width * scale * 0.5;
+        width = width * scale;
+        height = height * scale;
+
+        var bbox = Billboard.getScreenSpaceBoundingBox(b, Cartesian2.ZERO);
+        expect(bbox.x).toEqual(-halfWidth);
+        expect(bbox.y).toEqual(0);
+        expect(bbox.width).toEqual(width);
+        expect(bbox.height).toEqual(height);
+
+        b.verticalOrigin = VerticalOrigin.TOP;
+        bbox = Billboard.getScreenSpaceBoundingBox(b, Cartesian2.ZERO);
+        expect(bbox.x).toEqual(-halfWidth);
+        expect(bbox.y).toEqual(-height);
+        expect(bbox.width).toEqual(width);
+        expect(bbox.height).toEqual(height);
+    });
+
+    it('computes screen space bounding box with horizontal origin', function() {
+        var width = 10;
+        var height = 15;
+        var scale = 1.5;
+
+        var b = billboards.add({
+            width : width,
+            height : height,
+            scale : scale,
+            horizontalOrigin : HorizontalOrigin.LEFT
+        });
+
+        var halfHeight = height * scale * 0.5;
+        height = height * scale;
+        width = width * scale;
+
+        var bbox = Billboard.getScreenSpaceBoundingBox(b, Cartesian2.ZERO);
+        expect(bbox.x).toEqual(0);
+        expect(bbox.y).toEqual(-halfHeight);
+        expect(bbox.width).toEqual(width);
+        expect(bbox.height).toEqual(height);
+
+        b.horizontalOrigin = HorizontalOrigin.RIGHT;
+        bbox = Billboard.getScreenSpaceBoundingBox(b, Cartesian2.ZERO);
+        expect(bbox.x).toEqual(-width);
+        expect(bbox.y).toEqual(-halfHeight);
+        expect(bbox.width).toEqual(width);
+        expect(bbox.height).toEqual(height);
     });
 
     it('equals another billboard', function() {
@@ -1504,10 +1614,10 @@ defineSuite([
 
         // switch to blue while green is in-flight
 
-        one.image = './Data/Images/Blue.png';
+        one.image = './Data/Images/Blue10x10.png';
 
         expect(one.ready).toEqual(false);
-        expect(one.image).toEqual('./Data/Images/Blue.png');
+        expect(one.image).toEqual('./Data/Images/Blue10x10.png');
 
         return pollToPromise(function() {
             return one.ready;